[Refactor]: Change network_blocks to trasformer_blocks
