+++ mmselfsup/models/utils/__init__.py
@@ -6,8 +6,8 @@ from .gather_layer import GatherLayer
 from .knn_classifier import knn_classifier
 from .multi_pooling import MultiPooling
 from .multi_prototypes import MultiPrototypes
-from .network_blocks import (CAETransformerRegressorLayer, MultiheadAttention,
-                             TransformerEncoderLayer)
+from .transformer_blocks import (CAETransformerRegressorLayer,
+                                 MultiheadAttention, TransformerEncoderLayer)
 from .position_embedding import build_2d_sincos_position_embedding
 from .sobel import Sobel
